video directory now builds under windows.
svn path=/trunk/libavg/; revision=2378

+++ src/video/AsyncDemuxer.h
@@ -39,7 +39,7 @@ namespace avg {
            
             void enableStream(int StreamIndex);
             AVPacket * getPacket(int StreamIndex);
-            void seek(int DestFrame, int StartTimestamp, int StreamIndex);
+            void seek(int DestFrame, long long StartTimestamp, int StreamIndex);
             
         private:
             void waitForSeekDone();
